Adds image map to style.
This is used to map different types of inset sections to the correct
images used to style them. I would just use the css background stuff and
delegate this task to the stylesheets, but browsers don't know really
how to handle css backgrounds of svgs when svgs are served locally.