Adds chapter index option and uses it to mark figures and tables.
Chapters get an :index property that takes a string that is prepended to
figure captions to mark them as part of the book as a whole. Chapter 1
will have an index of "1" by default through the Book class, which means
figures are 1-1, 1-2, etc.

Added references of the form @[tag] where tag is given by the figure or
table and then can be used anywhere throughout the chapter.

Images use this syntax to provide a tag (named 'tag'):

![Image Alt](tag@file.png "Title Text")

Tables use this syntax to provide a tag (named 'tag'):

!table tag "Table title"