Fixes local style resources and adds local style config file.
Big commit due to lots of breakage.

Adds a style_config field to the book yaml that allows you to override
the default location it looks for the style config. Defaults to
style.yml.

It reads the style.yml which merely contains the list of local assets in
the order they should be included in the document.

To faciliate this, Style and Asset need to differentiate between
style/gem supplied assets and local assets more effectively. So, a
style_supplied? method is added to Asset and the style can be passed
into Asset to provide the asset path.